Renamed package model
