+++ src/com/liquidbol/db/persistence/ExpenseCrud.java
@@ -6,8 +6,8 @@
 
 package com.liquidbol.db.persistence;
 
-import com.liquidbol.commons.model.Expense;
-import com.liquidbol.commons.model.Store;
+import com.liquidbol.model.Expense;
+import com.liquidbol.model.Store;
 import java.sql.Connection;
 import java.sql.Date;
 import java.sql.PreparedStatement;
