+++ src/com/liquidbol/db/persistence/ItemCrud.java
@@ -6,7 +6,7 @@
 
 package com.liquidbol.db.persistence;
 
-import com.liquidbol.commons.model.Item;
+import com.liquidbol.model.Item;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
