+++ src/com/liquidbol/db/persistence/ItemPaymentCrud.java
@@ -6,9 +6,9 @@
 
 package com.liquidbol.db.persistence;
 
-import com.liquidbol.commons.model.BillPayment;
-import com.liquidbol.commons.model.Employee;
-import com.liquidbol.commons.model.ItemBill;
+import com.liquidbol.model.BillPayment;
+import com.liquidbol.model.Employee;
+import com.liquidbol.model.ItemBill;
 import java.sql.Connection;
 import java.sql.Date;
 import java.sql.PreparedStatement;
