+++ src/com/liquidbol/db/persistence/ItemSaleCrud.java
@@ -6,9 +6,9 @@
 
 package com.liquidbol.db.persistence;
 
-import com.liquidbol.commons.model.Item;
-import com.liquidbol.commons.model.ItemBill;
-import com.liquidbol.commons.model.ItemSale;
+import com.liquidbol.model.Item;
+import com.liquidbol.model.ItemBill;
+import com.liquidbol.model.ItemSale;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
