+++ src/com/liquidbol/db/persistence/ServiceBillCrud.java
@@ -6,9 +6,9 @@
 
 package com.liquidbol.db.persistence;
 
-import com.liquidbol.commons.model.Client;
-import com.liquidbol.commons.model.Employee;
-import com.liquidbol.commons.model.ServiceBill;
+import com.liquidbol.model.Client;
+import com.liquidbol.model.Employee;
+import com.liquidbol.model.ServiceBill;
 import java.sql.Connection;
 import java.sql.Date;
 import java.sql.PreparedStatement;
