+++ src/com/liquidbol/db/persistence/ServiceReceptionCrud.java
@@ -6,10 +6,10 @@
 
 package com.liquidbol.db.persistence;
 
-import com.liquidbol.commons.model.RechargeableItem;
-import com.liquidbol.commons.model.Service;
-import com.liquidbol.commons.model.ServiceBill;
-import com.liquidbol.commons.model.ServiceReception;
+import com.liquidbol.model.RechargeableItem;
+import com.liquidbol.model.Service;
+import com.liquidbol.model.ServiceBill;
+import com.liquidbol.model.ServiceReception;
 import java.sql.Connection;
 import java.sql.Date;
 import java.sql.PreparedStatement;
