+++ src/com/liquidbol/db/persistence/StoreCrud.java
@@ -6,7 +6,7 @@
 
 package com.liquidbol.db.persistence;
 
-import com.liquidbol.commons.model.Store;
+import com.liquidbol.model.Store;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
