+++ src/com/liquidbol/model/ItemSale.java
@@ -4,7 +4,7 @@
  * and open the template in the editor.
  */
 
-package com.liquidbol.commons.model;
+package com.liquidbol.model;
 
 import com.liquidbol.db.persistence.ItemCrud;
 import com.liquidbol.db.persistence.PersistenceException;
@@ -13,26 +13,29 @@ import java.util.logging.Level;
 import java.util.logging.Logger;
 
 /**
- * Class that represents an item request.
+ * Class that represents an item sale.
  * @author Allan Leon
  */
-public class ItemRequest implements Serializable {
+public class ItemSale implements Serializable {
     
     private int id;
     private Item item;
     private int quantity;
     private Double amount;
+    private String obs;
 
     /**
      * Constructor method.
      * @param id
      * @param item
      * @param quantity
+     * @param obs 
      */
-    public ItemRequest(int id, Item item, int quantity) {
+    public ItemSale(int id, Item item, int quantity, String obs) {
         this.id = id;
         this.item = item;
         this.quantity = quantity;
+        this.obs = obs;
         this.amount = item.getPrice() * quantity;
     }
 
@@ -41,13 +44,15 @@ public class ItemRequest implements Serializable {
      * @param id
      * @param item
      * @param quantity
-     * @param amount 
+     * @param amount
+     * @param obs 
      */
-    public ItemRequest(int id, Item item, int quantity, Double amount) {
+    public ItemSale(int id, Item item, int quantity, Double amount, String obs) {
         this.id = id;
         this.item = item;
         this.quantity = quantity;
         this.amount = amount;
+        this.obs = obs;
     }
 
     /**
@@ -79,6 +84,13 @@ public class ItemRequest implements Serializable {
     }
 
     /**
+     * @return the obs
+     */
+    public String getObs() {
+        return obs;
+    }
+
+    /**
      * @param id the id to set
      */
     public void setId(int id) {
@@ -105,6 +117,13 @@ public class ItemRequest implements Serializable {
     public void setAmount(Double amount) {
         this.amount = amount;
     }
+
+    /**
+     * @param obs the obs to set
+     */
+    public void setObs(String obs) {
+        this.obs = obs;
+    }
     
     public void refresh() {
         try {
@@ -118,7 +137,7 @@ public class ItemRequest implements Serializable {
 
     @Override
     public int hashCode() {
-        int hash = 7;
+        int hash = 3;
         hash = 47 * hash + this.id;
         return hash;
     }
@@ -131,7 +150,7 @@ public class ItemRequest implements Serializable {
         if (getClass() != obj.getClass()) {
             return false;
         }
-        final ItemRequest other = (ItemRequest) obj;
+        final ItemSale other = (ItemSale) obj;
         if (this.id != other.id) {
             return false;
         }
